git.sh: stop touching global git config. Fixes: #7907
The Armbian project should not touch a user's git configuration outside
of project folders.  This commit fixes that.

The current solution is still quite a bit of a hack as we are opting
out of a safety feature on the users' behalf without their consent.

As of now, we add the full and absolute path to the list of safe
directories. If the directory is moved, it will need to be readded.
It might be preferrable to use '.' instead for the relative path. If
we can limit opting out of the entire working directory but use only a
subset of subdirectories (like ./build or ./sources), that would be an
improvement which ought to be made in the future.

Closing #7907 AR-2624